requirements.txt
absl-py==0.8.0
astor==0.8.0
CacheControl==0.12.5
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
future==0.17.1
gast==0.3.1
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.10.0
idna==2.8
imageio==2.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
msgpack==0.6.1
networkx==2.3
numpy==1.17.2
opencv-python==3.4.7.28
openslide-python==1.1.1
pandas==0.25.1
Pillow==6.1.0
protobuf==3.9.1
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyWavelets==1.0.3
requests==2.22.0
requests-toolbelt==0.9.1
scikit-image==0.15.0
scipy==1.2.0
seaborn==0.9.0
Shapely==1.6.4.post2
six==1.12.0
tensorflow==1.14.0
